+++ bundlers/browser/BrowserMultiPack.ts
@@ -82,6 +82,49 @@ module BrowserMultiPack {
     }
 
 
+    /** Creates a custom 'browser-pack' implementation that writes to multiple output streams ('bundles').
+     * @param getMultiBundleOpts a function which returns a MultiBundleOptions object containing the options to build the bundle streams.
+     * This function gets called when browserify.bundle() is called, which happens in BundleBuilder.compileBundle() (which calls BrowserifyHelper.setupRebundleListener())
+     * @returns two functions: 'createSplitPacker()' which builds a 'browser-pack' like stream and 'multiBundleSourceCreator()' which is meant to be passed to bundleBuilder.setBundleSourceCreator()'
+     */
+    export function createPacker(getMultiBundleOpts: () => MultiBundleOptions) {
+        var inst = {
+            pack: <{ baseStream: stream.Transform; bundleStreams: BundleStream<stream.Transform>[] } | null>null,
+            updateDeps: <string[] | null>null,
+
+            // gets called when browser bundler instance is created or when reset() or bundle() are called
+            createPackStreams: function createPackStreams() {
+                var multiBundleOpts = getMultiBundleOpts();
+                var streamsToUpdate = bundlesToUpdate(multiBundleOpts, inst.updateDeps);
+                inst.pack = BrowserMultiPack.createPackStreams(multiBundleOpts, streamsToUpdate);
+
+                return inst.pack;
+            },
+
+            // Consume the browser bundle and return the multiple pack bundles
+            multiBundleSourceCreator: function multiBundleSourceCreator<TBundler extends { bundle(): NodeJS.ReadableStream }>(bundler: TBundler, updateEvent?: { [key: string]: any }) {
+                if (updateEvent != null) {
+                    inst.updateDeps = Object.keys(updateEvent).map((k) => updateEvent[k]);
+                }
+                var brBundle = bundler.bundle();
+                var res: any[] = [];
+                // When the bundle stream has available data, read it so that the stream ends
+                brBundle.on("readable", () => {
+                    var rec;
+                    while ((rec = brBundle.read()) !== null) {
+                        res.push(rec);
+                    }
+                });
+                return {
+                    baseStream: brBundle,
+                    bundleStreams: (<Exclude<typeof inst["pack"], null>>inst.pack).bundleStreams
+                };
+            }
+        };
+        return inst;
+    }
+
+
     /** Override browserify's standard 'pack' pipeline step with a custom 'browser-pack' implementation that writes to multiple output bundles.
      * This requires overwriting browserif.prototype._createPipeline() and setting the 'bundleBldr' setBundleSourceCreator() callback
      * @param bundleBldr the bundle builder to modify
@@ -95,8 +138,8 @@ module BrowserMultiPack {
         getMultiBundleOpts: () => MultiBundleOptions
     ) {
         var origCreatePipeline = _bundler.prototype["_createPipeline"];
-        var newBpack: { baseStream: stream.Transform; bundleStreams: BundleStream<stream.Transform>[] };
-        var updateDeps: string[] | null = null;
+
+        var packer = createPacker(getMultiBundleOpts);
 
         // Override browserify._createPipeline() to replace the 'pack' pipeline step with a custom browser-pack implementation
         // gets called when browserify instance is created or when reset() or bundle() are called
@@ -105,9 +148,7 @@ module BrowserMultiPack {
             var packPipe = pipeline.get("pack");
             var oldBpack = packPipe.pop();
 
-            var multiBundleOpts = getMultiBundleOpts();
-            var streamsToUpdate = bundlesToUpdate(multiBundleOpts, updateDeps);
-            newBpack = BrowserMultiPack.createPackStreams(multiBundleOpts, streamsToUpdate);
+            var newBpack = packer.createPackStreams();
 
             this._bpack = newBpack.baseStream;
             packPipe.push(newBpack.baseStream);
@@ -115,24 +156,9 @@ module BrowserMultiPack {
         };
 
         // Consume the browserify bundle and return the multiple pack bundles
-        bundleBldr.setBundleSourceCreator(function multiBundleStreamCreator(bundler: TBundler, updateEvent?: { [key: string]: any }) {
-            if (updateEvent != null) {
-                updateDeps = Object.keys(updateEvent).map((k) => updateEvent[k]);
-            }
-            var brwsBundle = bundler.bundle();
-            var res: any[] = [];
-            // When the bundle stream has available data, read it so that the stream ends
-            brwsBundle.on("readable", () => {
-                var rec;
-                while ((rec = brwsBundle.read()) !== null) {
-                    res.push(rec);
-                }
-            });
-            return {
-                baseStream: brwsBundle,
-                bundleStreams: newBpack.bundleStreams
-            };
-        });
+        bundleBldr.setBundleSourceCreator(packer.multiBundleSourceCreator);
+
+        return packer;
     }